kafka_vs_mqtt/docker_image/requirements.txt
lz4
crc32c
pandas
matplotlib
objsize
paho-mqtt